llm_service/requirements.txt
fastapi
uvicorn
openai
weaviate-client
